// src/app/commons/models/basic-information.ts
export class BasicInformation {
  id?: string;
  region?: string;
  country = {name: 'Estados Unidos', alpha2Code: 'US', alpha3Code: 'USA', numericCode: '840', callingCode: '+1'};
  city?: string;
  business?: string;
  webPage?: string;
  projectLeaderName?: string;
  emailProjectLeader?: string;
  cellProjectLeader?: string;
  collaborationSiteLink?: string;
  industry?: string;
  mainSegments?: string;
  mainCompetitors?: any;
  numEmployees?: string;
  projectStatus?: string;
  estimatedSales?: any;
  marketingBudget?: any;
}